Added a class testing the fit between a simulated EBSD pattern and a GeometricalKikuchiPatternSimulation
The class TestFitPatternDetectorOrientation has been added to
tests/test_signals/test_ebsd_master_pattern.py. The class contains
the test test_fit_detector_orientation(), which is parametrised with
8 sets of detector tilt, azimuthal and euler_2 angles, and lists
of associated zone axes, which are on the pattern simulated under
those conditions. For each set of parameters, we expect that the
fit between the simulated pattern and the GeometricalKikuchiPatternSimulation
is perfect.

These tests are currently all passing.

Signed-off-by: tgwoodcock <tgwoodcock@users.noreply.github.com>