Updated tests calling _get_direction_cosines_for_fixed_pc() and _get_direction_cosines_for_varying_pc()
In tests/test_rotation/test_rotation.py:
In TestRotationVectorTools.test_rotate_vector(), added the instantiation
of an EBSDDetector so that the gnomonic_bounds can be passed into the
new signature of _get_direction_cosines_for_fixed_pc(). Modified the
arguments in the function call accordingly.

In tests/test_signals/test_ebsd_master_pattern.py:
In TestProjectFromLambert.test_get_direction_cosines(), updated the
arguments in the function call to _get_direction_cosines_for_fixed_pc().
Added a further call to that function but WITHOUT .py_func and added a
further assert statement to check that the results with and without
.py_func are the same (as recommended in the contributing guidelines).

In TestProjectFromLambert.test_get_direction_cosines_for_multiple_pcs(),
updated the arguments in the function call to
_get_direction_cosines_for_varying_pc().
Added a further call to that function but WITHOUT .py_func and added a
further assert statement to check that the results with and without
.py_func are the same (as recommended in the contributing guidelines).

In TestProjectFromLambert.test_detector_azimuthal()
As the function_get_direction_cosines_for_fixed_pc() has been
modified, the small rotation in the EBSD pattern simulated from
the master pattern for tilt!=0 and azimuthal!=0 is no longer
present. The mean values of the data in the simulated patterns
sim2 and sim3 in the above test are therefore very slightly
different to the old values. The mean values have been updated.

The above tests are now all passing.

Signed-off-by: tgwoodcock <tgwoodcock@users.noreply.github.com>