fix(GeometricalKikuchiPatternSimulation): plotting with gnomonic coordinates
Problem: calling GeometricalKikuchiPatternSimulation.plot() with the kwarg
coordinates="gnomonic" and an EBSDDetector with navigation_dimension > 1 fails
with a ValueError.

Details: plot() calls as_collections(), which calls _zone_axes_as_collection()
and _zone_axis_labels_as_list(). The ValueError comes from the definition of
scatter_size in the "else" branch of _zone_axes_as_collection()
AND because of the application of the y_offset to the "coords" in the
"else" branch of _zone_axis_labels_as_list(). For an EBSDDetector with
navigation_dimension > 1, the call np.diff(self.detector.x_range)[0] results
in a 2D output, whereas a float is needed to calculate the offsets,

Solution: add the "index" to each of these lines so that the specific x_range
and y_range for the pattern specified by "index" is chosen. Leave the "[0]" there
so that we select a float rather than a np.ndarray. The case that a 1d detector is
supplied together with a >1d rotation is also handled.

Signed-off-by: tgwoodcock <tgwoodcock@users.noreply.github.com>