fix(_get_lambert_interpolation_parameters): avoid out-of-bounds values
The outputs nii, nij, niip and nijp of _get_lambert_interpolation_parameters()
are used by the function _get_pixel_from_master_pattern() to select a pixel.

In order to avoid IndexError, nii and niip must be in the range
0 <= n < npy and nij and nijp must be in the range
0 <= n < npx. The outputs nii and nij are automatically in these
ranges; however, niip and nijp may not be and we need to *exclude*
indices which are < 0 or >= npx or >= npy.

Previously, _get_lambert_interpolation_parameters() checked if
niip and nijp were > npx or > npy. This has now been changed to
>= npx and >= npy to fulfil the above conditions. This function
now only outputs indices which are in range for the master pattern
and therefore no further problems in _get_pixel_from_master_pattern()
are caused.

Signed-off-by: tgwoodcock <tgwoodcock@users.noreply.github.com>